chore: Cleaning up unused code
