+++ internal/git/git_test.go
@@ -19,15 +19,8 @@ limitations under the License.
 package git
 
 import (
-	"context"
-	"os"
-	"path/filepath"
 	"testing"
-	"time"
 
-	"github.com/go-git/go-git/v5"
-	"github.com/go-git/go-git/v5/plumbing"
-	"github.com/go-git/go-git/v5/plumbing/object"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
@@ -440,75 +433,6 @@ func TestGetAuthMethod_EmptyKey(t *testing.T) {
 	assert.Nil(t, auth)
 }
 
-func TestClone_BasicCall(t *testing.T) {
-	// Test that Clone properly handles invalid repository URLs
-	// Since we're using a fake URL, we expect this to fail
-	repo, err := Clone("https://github.com/test/repo.git", "/tmp/test", nil)
-
-	require.Error(t, err) // Expect error with fake repository
-	assert.Nil(t, repo)   // Repo should be nil on error
-}
-
-func TestRepo_Commit_BasicCall(t *testing.T) {
-	// Since Commit requires a real git repository, we expect this to fail
-	// when trying to call Worktree() on a nil Repository
-	repo := &Repo{path: "/tmp/test"}
-
-	files := []CommitFile{
-		{
-			Path:    "test/file1.yaml",
-			Content: []byte("apiVersion: v1\nkind: Pod"),
-		},
-		{
-			Path:    "test/file2.yaml",
-			Content: []byte("apiVersion: v1\nkind: Service"),
-		},
-	}
-
-	message := "Test commit message"
-
-	// Expect error since Repository is nil
-	err := repo.Commit(files, message)
-	require.Error(t, err)
-}
-
-func TestRepo_Commit_EmptyFiles(t *testing.T) {
-	// Since Commit requires a real git repository, we expect this to fail
-	repo := &Repo{path: "/tmp/test"}
-
-	var files []CommitFile
-	message := "Empty commit"
-
-	// Expect error since Repository is nil
-	err := repo.Commit(files, message)
-	require.Error(t, err)
-}
-
-func TestRepo_Commit_EmptyMessage(t *testing.T) {
-	// Since Commit requires a real git repository, we expect this to fail
-	repo := &Repo{path: "/tmp/test"}
-
-	files := []CommitFile{
-		{
-			Path:    "test/file.yaml",
-			Content: []byte("content"),
-		},
-	}
-
-	// Expect error since Repository is nil
-	err := repo.Commit(files, "")
-	require.Error(t, err)
-}
-
-func TestRepo_Push_BasicCall(t *testing.T) {
-	// Since Push requires a real git repository, we expect this to fail
-	repo := &Repo{path: "/tmp/test"}
-
-	// Expect error since Repository is nil
-	err := repo.Push(context.Background())
-	require.Error(t, err)
-}
-
 func TestCommitFile_Structure(t *testing.T) {
 	file := CommitFile{
 		Path:    "namespaces/default/Pod/test-pod.yaml",
@@ -868,336 +792,4 @@ func TestBatchOperations_MultipleDeletes(t *testing.T) {
 
 	// Verify we have the expected number of events
 	assert.Len(t, events, 4)
-}
-
-func TestCheckout_BranchCreationOnEmptyRepo(t *testing.T) {
-	// Test that Checkout creates branches automatically on empty repositories
-	tempDir := t.TempDir()
-
-	// Initialize an empty repository (simulating what initializeEmptyRepository does)
-	repo, err := git.PlainInit(tempDir, false)
-	require.NoError(t, err)
-
-	// Create initial commit
-	worktree, err := repo.Worktree()
-	require.NoError(t, err)
-
-	gitkeepPath := filepath.Join(tempDir, ".gitkeep")
-	err = os.WriteFile(gitkeepPath, []byte("# This file ensures the repository is not empty\n"), 0600)
-	require.NoError(t, err)
-
-	_, err = worktree.Add(".gitkeep")
-	require.NoError(t, err)
-
-	_, err = worktree.Commit("Initial commit", &git.CommitOptions{
-		Author: &object.Signature{
-			Name:  "GitOps Reverser",
-			Email: "gitops-reverser@configbutler.ai",
-			When:  time.Now(),
-		},
-	})
-	require.NoError(t, err)
-
-	// Create a Repo wrapper
-	gitRepo := &Repo{
-		Repository: repo,
-		path:       tempDir,
-		branch:     "main",
-		remoteName: "origin",
-	}
-
-	// Test checking out a new branch that doesn't exist
-	newBranch := "test-feature-branch"
-	err = gitRepo.Checkout(newBranch)
-	require.NoError(t, err)
-
-	// Verify the branch was created and we're on it
-	head, err := repo.Head()
-	require.NoError(t, err)
-	assert.Equal(t, plumbing.NewBranchReferenceName(newBranch), head.Name())
-
-	// Verify we can switch back to main
-	err = gitRepo.Checkout("main")
-	require.NoError(t, err)
-
-	head, err = repo.Head()
-	require.NoError(t, err)
-	assert.Equal(t, plumbing.NewBranchReferenceName("main"), head.Name())
-}
-
-func TestCheckout_BranchDoesNotExist(t *testing.T) {
-	// Test that Checkout creates branches automatically when they don't exist
-	tempDir := t.TempDir()
-
-	// Initialize repository with initial commit
-	repo, err := git.PlainInit(tempDir, false)
-	require.NoError(t, err)
-
-	worktree, err := repo.Worktree()
-	require.NoError(t, err)
-
-	gitkeepPath := filepath.Join(tempDir, ".gitkeep")
-	err = os.WriteFile(gitkeepPath, []byte("# Initial commit\n"), 0600)
-	require.NoError(t, err)
-
-	_, err = worktree.Add(".gitkeep")
-	require.NoError(t, err)
-
-	_, err = worktree.Commit("Initial commit", &git.CommitOptions{
-		Author: &object.Signature{
-			Name:  "GitOps Reverser",
-			Email: "gitops-reverser@configbutler.ai",
-			When:  time.Now(),
-		},
-	})
-	require.NoError(t, err)
-
-	// Create a Repo wrapper
-	gitRepo := &Repo{
-		Repository: repo,
-		path:       tempDir,
-		branch:     "main",
-		remoteName: "origin",
-	}
-
-	// Try to checkout a branch that doesn't exist
-	// This should work because our Checkout method creates branches automatically
-	err = gitRepo.Checkout("nonexistent-branch")
-	require.NoError(t, err)
-
-	// Verify we're on the new branch
-	head, err := repo.Head()
-	require.NoError(t, err)
-	assert.Equal(t, plumbing.NewBranchReferenceName("nonexistent-branch"), head.Name())
-}
-
-func TestClone_EmptyRepositoryHandling(t *testing.T) {
-	// Test that Clone handles empty repositories by initializing local repo
-	tempDir := t.TempDir()
-
-	// Create a mock empty repository by initializing a bare repo
-	emptyRepoDir := filepath.Join(tempDir, "empty-repo")
-	err := os.MkdirAll(emptyRepoDir, 0750)
-	require.NoError(t, err)
-
-	// Initialize as bare repository (simulating an empty remote)
-	_, err = git.PlainInit(emptyRepoDir, true) // true = bare
-	require.NoError(t, err)
-
-	// Try to clone from this "empty" repository
-	// Since it's a local path, we need to use file:// protocol
-	repoURL := "file://" + emptyRepoDir
-	clonePath := filepath.Join(tempDir, "cloned-repo")
-
-	// This should succeed because our Clone function handles empty repos
-	repo, err := Clone(repoURL, clonePath, nil)
-	require.NoError(t, err) // Should succeed due to empty repo handling
-	require.NotNil(t, repo)
-
-	// Verify the repository was initialized (HEAD should not exist for empty repo)
-	_, err = repo.Head()
-	// For an empty repository, HEAD should not exist yet
-	require.Error(t, err) // Expect error because no commits exist
-
-	// Check that we have no commits (Log should fail on empty repo)
-	commits, err := repo.Log(&git.LogOptions{})
-	require.Error(t, err) // Log should fail on empty repository
-	assert.Nil(t, commits)
-}
-
-func TestGenerateLocalCommits_FirstCommitOnEmptyRepo(t *testing.T) {
-	// Test that the system can create the first Kubernetes resource commit
-	// after repository initialization (lazy initialization scenario)
-	tempDir := t.TempDir()
-
-	// Initialize repository with minimal .gitkeep commit (standard for empty repos)
-	repo, err := git.PlainInit(tempDir, false)
-	require.NoError(t, err)
-
-	worktree, err := repo.Worktree()
-	require.NoError(t, err)
-
-	// Create minimal initial commit (standard practice for new repos)
-	gitkeepPath := filepath.Join(tempDir, ".gitkeep")
-	err = os.WriteFile(gitkeepPath, []byte(""), 0600)
-	require.NoError(t, err)
-
-	_, err = worktree.Add(".gitkeep")
-	require.NoError(t, err)
-
-	_, err = worktree.Commit("Initial commit", &git.CommitOptions{
-		Author: &object.Signature{
-			Name:  "GitOps Reverser",
-			Email: "gitops-reverser@configbutler.ai",
-			When:  time.Now(),
-		},
-	})
-	require.NoError(t, err)
-
-	// Create a Repo wrapper
-	gitRepo := &Repo{
-		Repository: repo,
-		path:       tempDir,
-		branch:     "main",
-		remoteName: "origin",
-	}
-
-	// Create first event
-	obj := &unstructured.Unstructured{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "ConfigMap",
-			"metadata": map[string]interface{}{
-				"name":            "first-config",
-				"namespace":       "default",
-				"resourceVersion": "1",
-			},
-			"data": map[string]interface{}{
-				"key": "value",
-			},
-		},
-	}
-
-	event := Event{
-		Object: obj,
-		Identifier: types.ResourceIdentifier{
-			Group:     "",
-			Version:   "v1",
-			Resource:  "configmaps",
-			Namespace: "default",
-			Name:      "first-config",
-		},
-		Operation: "CREATE",
-		UserInfo: UserInfo{
-			Username: "admin",
-			UID:      "admin-uid",
-		},
-		BaseFolder: "",
-	}
-
-	ctx := context.Background()
-
-	// Generate local commits (should create the first commit)
-	commitsCreated, err := gitRepo.generateLocalCommits(ctx, []Event{event})
-	require.NoError(t, err, "Should successfully create first commit on empty repo")
-	assert.Equal(t, 1, commitsCreated, "Should create exactly 1 commit")
-
-	// Verify repository now has a commit
-	head, err := repo.Head()
-	require.NoError(t, err, "Repository should now have HEAD after first commit")
-	assert.NotNil(t, head)
-
-	// Verify file was created
-	filePath := filepath.Join(tempDir, "v1/configmaps/default/first-config.yaml")
-	assert.FileExists(t, filePath, "File should exist after commit")
-
-	// Verify file content
-	content, err := os.ReadFile(filePath)
-	require.NoError(t, err)
-	assert.Contains(t, string(content), "name: first-config")
-	assert.Contains(t, string(content), "namespace: default")
-	assert.Contains(t, string(content), "key: value")
-
-	// Verify commit message
-	commitObj, err := repo.CommitObject(head.Hash())
-	require.NoError(t, err)
-	assert.Equal(t, "[CREATE] v1/configmaps/first-config by user/admin", commitObj.Message)
-
-	// Verify we're on the main branch
-	assert.Equal(t, plumbing.NewBranchReferenceName("main"), head.Name())
-}
-
-func TestTryPushCommits_FirstCommitOnEmptyRepoWithBranchCreation(t *testing.T) {
-	// Test the complete flow: initialized repo + first event + branch creation + commit + push
-	tempDir := t.TempDir()
-
-	// Create "remote" repository with initial commit
-	remoteRepoPath := filepath.Join(tempDir, "remote")
-	remoteRepo, err := git.PlainInit(remoteRepoPath, false) // non-bare for easier testing
-	require.NoError(t, err)
-
-	// Create initial commit in remote
-	err = createInitialCommitHelper(remoteRepo, remoteRepoPath)
-	require.NoError(t, err)
-
-	// Clone to local
-	localRepoPath := filepath.Join(tempDir, "local")
-	localRepo, err := git.PlainClone(localRepoPath, false, &git.CloneOptions{
-		URL: remoteRepoPath,
-	})
-	require.NoError(t, err)
-
-	gitRepo := &Repo{
-		Repository: localRepo,
-		path:       localRepoPath,
-		auth:       nil, // No auth for local test
-		branch:     "feature/first",
-		remoteName: "origin",
-	}
-
-	// Create first event
-	event := Event{
-		Object: createTestPodWithResourceVersion("initial-pod", "default", "1"),
-		Identifier: types.ResourceIdentifier{
-			Group:     "",
-			Version:   "v1",
-			Resource:  "pods",
-			Namespace: "default",
-			Name:      "initial-pod",
-		},
-		Operation: "CREATE",
-		UserInfo:  UserInfo{Username: "first-user", UID: "uid-1"},
-	}
-
-	ctx := context.Background()
-
-	// This should:
-	// 1. Create the branch "feature/first" (doesn't exist yet)
-	// 2. Create the first commit
-	// 3. Push to remote successfully
-	err = gitRepo.TryPushCommits(ctx, "feature/first", []Event{event})
-	require.NoError(t, err, "Should successfully create & push first commit on empty repo")
-
-	// Verify local state
-	head, err := localRepo.Head()
-	require.NoError(t, err, "Should have HEAD after first commit")
-	assert.Equal(t, plumbing.NewBranchReferenceName("feature/first"), head.Name())
-
-	// Verify file exists
-	filePath := filepath.Join(localRepoPath, "v1/pods/default/initial-pod.yaml")
-	assert.FileExists(t, filePath)
-
-	// Verify remote has the branch and commit
-	remoteBranchRef := plumbing.NewRemoteReferenceName("origin", "feature/first")
-	ref, err := localRepo.Reference(remoteBranchRef, true)
-	require.NoError(t, err, "Remote should have the pushed branch")
-	assert.Equal(t, head.Hash(), ref.Hash(), "Remote and local should have same commit")
-}
-
-// createInitialCommitHelper creates an initial commit in a repository for testing.
-func createInitialCommitHelper(repo *git.Repository, repoPath string) error {
-	worktree, err := repo.Worktree()
-	if err != nil {
-		return err
-	}
-
-	gitkeepPath := filepath.Join(repoPath, ".gitkeep")
-	err = os.WriteFile(gitkeepPath, []byte(""), 0600)
-	if err != nil {
-		return err
-	}
-
-	_, err = worktree.Add(".gitkeep")
-	if err != nil {
-		return err
-	}
-
-	_, err = worktree.Commit("Initial commit", &git.CommitOptions{
-		Author: &object.Signature{
-			Name:  "GitOps Reverser",
-			Email: "gitops-reverser@configbutler.ai",
-			When:  time.Now(),
-		},
-	})
-	return err
 }
