--- internal/git/test_helpers_test.go
@@ -1,51 +1,0 @@
-/*
-SPDX-License-Identifier: Apache-2.0
-
-Copyright 2025 ConfigButler
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package git
-
-import (
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-)
-
-//nolint:unparam // namespace parameter kept for API consistency even though tests use "default"
-func createTestPodWithResourceVersion(name, namespace, resourceVersion string) *unstructured.Unstructured {
-	pod := createTestPod(name, namespace)
-	pod.SetResourceVersion(resourceVersion)
-	return pod
-}
-
-func createTestPod(name, namespace string) *unstructured.Unstructured {
-	return &unstructured.Unstructured{
-		Object: map[string]interface{}{
-			"apiVersion": "v1",
-			"kind":       "Pod",
-			"metadata": map[string]interface{}{
-				"name":      name,
-				"namespace": namespace,
-			},
-			"spec": map[string]interface{}{
-				"containers": []interface{}{
-					map[string]interface{}{
-						"name":  "test-container",
-						"image": "nginx",
-					},
-				},
-			},
-		},
-	}
-}
